Refactor code to separate authentication code
