+++ requirements.txt
@@ -6,4 +6,4 @@ Flask-Script==2.0.5
 Flask-WTF==0.9.5
 simple-pbkdf2==1.0
 gunicorn==19.4
-pip==8.1.0
+pip==8.1.1
