+++ wolfstudy/__init__.py
@@ -2,7 +2,6 @@ from config import config
 from flask import Flask
 from flask.ext.bootstrap import Bootstrap
 from flask.ext.sqlalchemy import SQLAlchemy
-from flask.ext.wtf.csrf import CsrfProtect
 
 import errno
 import os
@@ -49,12 +48,14 @@ def create_app(config_name):
     app.config.from_object(config[config_name])
     config[config_name].init_app(app)
     install_secret_key(app)
-    CsrfProtect(app)
 
     bootstrap.init_app(app)
     db.init_app(app)
 
     from .main import main as main_blueprint
     app.register_blueprint(main_blueprint)
+
+    from .auth import auth as auth_blueprint
+    app.register_blueprint(auth_blueprint, url_prefix='/auth')
 
     return app
