+++ wolfstudy/main/views.py
@@ -1,7 +1,7 @@
-from flask import abort, flash, redirect, render_template, request, session, url_for
+from flask import flash, redirect, render_template, session, url_for
 from . import main
 from .. import db
-from .forms import AnswerQuestionForm, AskQuestionForm, LoginForm, RegisterForm
+from .forms import AnswerQuestionForm, AskQuestionForm
 from ..models import Answer, Question, User
 
 @main.route('/')
@@ -19,9 +19,8 @@ def get_question(question_id):
 @main.route('/question/<int:question_id>/answer/', methods=['POST'])
 def answer_question(question_id):
     if not session.get('logged_in', False):
-        print "not logged in"
         flash('You must be logged in to answer a question.')
-        return redirect(url_for('.login'))
+        return redirect(url_for('auth.login'))
 
     form = AnswerQuestionForm()
     if form.validate_on_submit():
@@ -42,7 +41,7 @@ def answer_question(question_id):
 def ask_question():
     if not session.get('logged_in', False):
         flash('You must be logged in to ask a question.')
-        return redirect(url_for('.login'))
+        return redirect(url_for('auth.login'))
 
     form = AskQuestionForm()
     if form.validate_on_submit():
@@ -52,64 +51,6 @@ def ask_question():
         return redirect(url_for('.get_question', question_id=new_question.id))
     else:
         return render_template('ask.html', form=form)
-
-@main.route('/register/', methods=['GET', 'POST'])
-def register():
-    form = RegisterForm()
-    if form.validate_on_submit():
-        email = form.email.data
-        username = form.username.data
-        password = form.password.data
-        password_retype = form.password_retype.data
-
-        error = None
-        if User.query.filter_by(username=username).first() != None:
-            error = 'Username already in use.'
-        elif User.query.filter_by(email=email).first() != None:
-            error = 'Email address already in use.'
-        elif password != password_retype:
-            error = 'The passwords you typed did not match.'
-
-        if error:
-            return render_template('register.html', error=error)
-
-        new_user = User(username, email, password)
-        db.session.add(new_user)
-        db.session.commit()
-
-        session['logged_in'] = True
-        session['username'] = username
-
-        return redirect(url_for('.index'))
-    else:
-        return render_template('register.html', form=form)
-
-@main.route('/login/', methods=['GET', 'POST'])
-def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-        username = form.username.data
-        password = form.password.data
-
-        user = User.query.filter_by(username=username).first()
-        
-        if user == None or not user.verify_password(password):
-            error = 'Login failed. Please try again.'
-            return render_template('login.html', form=form, error=error)
-
-        session['logged_in'] = True
-        session['username'] = username
-
-        return redirect(url_for('.index'))
-    else:
-        return render_template('login.html', form=form)
-
-@main.route('/logout/')
-def logout():
-    session['logged_in'] = False
-    session.pop('username')
-
-    return redirect(url_for('.index'))
 
 @main.route('/user/<username>/')
 def show_user(username):
