+++ wolfstudy/templates/base.html
@@ -8,11 +8,11 @@
     <meta name="description" content="">
 
     <!-- CSS -->
-    <link href="/static/css/normalize.css" rel="stylesheet">
-    <link href="/static/css/main.css" rel="stylesheet">
+    <link href="{{ url_for('static', filename='css/normalize.css') }}" rel="stylesheet">
+    <link href="{{ url_for('static', filename='css/main.css') }}" rel="stylesheet">
 
     <!-- JavaScript -->
-    <script src="/static/js/main.js" async></script>
+    <script src="{{ url_for('static', filename='js/main.js') }}" async></script>
 {% endblock %}
 
 {% block navbar %}
@@ -39,14 +39,14 @@
                 {% endif %}
 
                 <ul class="nav navbar-nav navbar-right">
-                    <li><a href="/">Home</a></li>
-                    <li><a href="/ask/">Ask</a></li>
+                    <li><a href="{{ url_for('main.index') }}">Home</a></li>
+                    <li><a href="{{ url_for('main.ask_question') }}">Ask</a></li>
 
                     {% if session.logged_in %}
-                        <li><a href="/logout/">Log out</a></li>
+                    <li><a href="{{ url_for('auth.logout') }}">Log out</a></li>
                     {% else %}
-                        <li><a href="/login/">Log in</a></li>
-                        <li><a href="/register/">Register</a></li>
+                    <li><a href="{{ url_for('auth.login') }}">Log in</a></li>
+                    <li><a href="{{ url_for('auth.register') }}">Register</a></li>
                     {% endif %}
                 </ul>
             </div>
