+++ wolfstudy/templates/index.html
@@ -8,7 +8,7 @@
 
         <p>{{ question.content }}</p>
 
-        <p><a href="/question/{{ question.id }}/">Read more...</a></p>
+        <p><a href="{{ url_for('main.get_question', question_id=question.id) }}">Read more...</a></p>
 
         <br>
         <br>
